moved action creator code in its own package
