+++ compiller/src/main/java/com/yheriatovych/reductor/processor/autoreducer/ReduceAction.java
@@ -5,7 +5,7 @@ import com.google.auto.common.MoreTypes;
 import com.yheriatovych.reductor.annotations.ActionCreator;
 import com.yheriatovych.reductor.annotations.AutoReducer;
 import com.yheriatovych.reductor.processor.*;
-import com.yheriatovych.reductor.processor.model.ActionCreatorElement;
+import com.yheriatovych.reductor.processor.actioncreator.ActionCreatorElement;
 
 import javax.lang.model.element.Element;
 import javax.lang.model.element.ExecutableElement;
